Decouple functions, more humane window treatment.
This update significantly decouples the functions that deal with
retrieving and displaying the forecasts, getting things set up for the
ability to have post-display hooks if users so desire.

Additionally, we no longer forcibly resize the display and instead
call (pop-to-buffer). Quitting Sunshine now calls (quit-window t ...)
rather than (kill-buffer), which leaves your windows in an expected
state.

Internally, we now save excursions at the appropriate times when
processing the icon downloads so that point remains at the beginning of
the buffer where you would expect it to be, instead of jumping all over
the place and usually ending up after the last icon position.