Force multibyte treatment of cache files
The `url-cache-extract' function forcibly disables multibyte in the
destination buffer and reads the file contents literally. This is great
for binary data, like images, but seems to destroy UTF-8-encoded JSON
text data.

Instead of calling `url-cache-extract', we now set multibyte on the
destination temporary buffer and read the cache file contents
"normally." This seems to fix #4.